ci(dev_env): Test development environment on Big Sur and Ubuntu
We have users out in the wild using `brew` on Linux, thus, we will serve our community
by testing the development environment there as well.

I'm also adding Big Sur back into the matrix (even if it's slower to execute) in order
to make sure we don't regress it when we make changes to the shared Brewfile.